Updated how the package handles overwriting of data.
It will actually set the environment-variables properly through
os.Setenv now.

It will always fall back to the last .env-file in the list. So if there
are two files in env.EnvironmentFiles, then it will start by setting and
using the first one, then go about and repeat this with the second file.
Never unsetting, but always overwriting with new data.